# services/datahub/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "DataHub manages datasources, documents, ingestions, and chunks.",
        "title": "DataHub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/documents/{document_id}/ingestions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestions"
                ],
                "summary": "List ingestions for a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.IngestionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestions"
                ],
                "summary": "Create an ingestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ingestion config",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateIngestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.IngestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/ingestions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingestions"
                ],
                "summary": "Get an ingestion by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingestion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.IngestionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ingestions"
                ],
                "summary": "Delete an ingestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ingestion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateIngestionRequest": {
            "type": "object",
            "properties": {
                "chunk_strategy": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "embedding_model": {
                    "type": "string"
                }
            }
        },
...
        "model.IngestionResponse": {
            "type": "object",
            "properties": {
                "chunk_strategy": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "embedding_model": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}